create skip sequencer module
