+++ scripts/sequitur.js
@@ -2,7 +2,8 @@ const fs = require("mz/fs"),
       { execSync } = require("child_process"),
       ef = require("./extractFrames"),
       cf = require('./compareFrames'),
-      sa = require('./sampleAudio');
+      sa = require('./sampleAudio'),
+      skip = require('./skipSequencer');
 
 console.clear();
 
@@ -43,43 +44,13 @@ if (!fs.existsSync(dir))
 (async() => {
     // load data
     const wave = await sa.wave(audioFile, framerate),
-          numOfFrames = await ef.frames(videoFile, resolution),
-          diffs = await cf.diffs(numOfFrames);
+          frames = await ef.frames(videoFile, resolution),
+          diffs = await cf.diffs(frames);
 
     // sequence and encode
-    const seq = sequence(numOfFrames, wave, framerate, maxLevel, minOffset);
+    const seq = skip.sequence(frames, wave, framerate, maxLevel, minOffset);
     encode(seq, resolution, framerate, audioFile, preview);
 })()
-
-function sequence(numOfFrames, wave, fps, max, min) {
-    let seq = [],
-        selectedFrame = Math.floor(numOfFrames * Math.random()),
-        frameDuration = 1/fps,
-        reverse = false;
-
-    for (let i = 0; i < (wave.length-1); i++) {
-        let level = wave[i],
-            maxLevel = max || 1,
-            minOffset = min || 0,
-            maxOffset = Math.round(numOfFrames * (level * maxLevel));
-        offset = maxOffset < minOffset ? 1 : maxOffset;
-
-        if ( selectedFrame + offset > numOfFrames-1 ) reverse = true;
-        else if ( selectedFrame - offset < 0 ) reverse = false;
-
-        if (reverse) selectedFrame = selectedFrame - offset;
-        else selectedFrame = selectedFrame + offset;
-
-        if ( selectedFrame > numOfFrames-1 || selectedFrame < 0 ) {
-            console.log('off the rails!');
-            process.exit();
-        }
-        
-        seq.push([selectedFrame, frameDuration]);
-    }
-
-    return seq;
-}
 
 function encode(seq, res, fps, aud, pre) {
     let seqStr = '';
