+++ scripts/skipSequencer.js
@@ -1,0 +1,29 @@
+exports.sequence = function(numOfFrames, wave, fps, max, min) {
+    let seq = [],
+        selectedFrame = Math.floor(numOfFrames * Math.random()),
+        frameDuration = 1/fps,
+        reverse = false;
+
+    for (let i = 0; i < (wave.length-1); i++) {
+        let level = wave[i],
+            maxLevel = max || 1,
+            minOffset = min || 0,
+            maxOffset = Math.round(numOfFrames * (level * maxLevel));
+        offset = maxOffset < minOffset ? 1 : maxOffset;
+
+        if ( selectedFrame + offset > numOfFrames-1 ) reverse = true;
+        else if ( selectedFrame - offset < 0 ) reverse = false;
+
+        if (reverse) selectedFrame = selectedFrame - offset;
+        else selectedFrame = selectedFrame + offset;
+
+        if ( selectedFrame > numOfFrames-1 || selectedFrame < 0 ) {
+            console.log('off the rails!');
+            process.exit();
+        }
+        
+        seq.push([selectedFrame, frameDuration]);
+    }
+
+    return seq;
+}
